+++ x/gov/genesis_test.go
@@ -34,7 +34,7 @@ func TestImportExportQueues(t *testing.T) {
 	ctx = app.BaseApp.NewContext(false, tmproto.Header{})
 
 	// Create two proposals, put the second into the voting period
-	proposal := TestProposal
+	proposal := govgenhelpers.TestTextProposal
 	proposal1, err := app.GovKeeper.SubmitProposal(ctx, proposal)
 	require.NoError(t, err)
 	proposalID1 := proposal1.ProposalId
@@ -90,7 +90,9 @@ func TestImportExportQueues(t *testing.T) {
 	ctx2 := app2.BaseApp.NewContext(false, tmproto.Header{})
 
 	// Jump the time forward past the DepositPeriod and VotingPeriod
-	ctx2 = ctx2.WithBlockTime(ctx2.BlockHeader().Time.Add(app2.GovKeeper.GetDepositParams(ctx2).MaxDepositPeriod).Add(app2.GovKeeper.GetVotingParams(ctx2).VotingPeriod))
+	ctx2 = ctx2.WithBlockTime(ctx2.BlockHeader().Time.
+		Add(app2.GovKeeper.GetDepositParams(ctx2).MaxDepositPeriod).
+		Add(app2.GovKeeper.GetVotingPeriod(ctx2, proposal)))
 
 	// Make sure that they are still in the DepositPeriod and VotingPeriod respectively
 	proposal1, ok = app2.GovKeeper.GetProposal(ctx2, proposalID1)
@@ -146,7 +148,7 @@ func TestEqualProposals(t *testing.T) {
 	app.BeginBlock(abci.RequestBeginBlock{Header: header})
 
 	// Submit two proposals
-	proposal := TestProposal
+	proposal := govgenhelpers.TestTextProposal
 	proposal1, err := app.GovKeeper.SubmitProposal(ctx, proposal)
 	require.NoError(t, err)