+++ x/gov/keeper/proposal_test.go
@@ -7,12 +7,14 @@ import (
 	"time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
 
+	govgenhelpers "github.com/atomone-hub/govgen/app/helpers"
 	"github.com/atomone-hub/govgen/x/gov/types"
 )
 
 func (suite *KeeperTestSuite) TestGetSetProposal() {
-	tp := TestProposal
+	tp := govgenhelpers.TestTextProposal
 	proposal, err := suite.app.GovKeeper.SubmitProposal(suite.ctx, tp)
 	suite.Require().NoError(err)
 	proposalID := proposal.ProposalId
@@ -24,7 +26,7 @@ func (suite *KeeperTestSuite) TestGetSetProposal() {
 }
 
 func (suite *KeeperTestSuite) TestActivateVotingPeriod() {
-	tp := TestProposal
+	tp := govgenhelpers.TestTextProposal
 	proposal, err := suite.app.GovKeeper.SubmitProposal(suite.ctx, tp)
 	suite.Require().NoError(err)
 
@@ -78,7 +80,7 @@ func (suite *KeeperTestSuite) TestGetProposalsFiltered() {
 
 	for _, s := range status {
 		for i := 0; i < 50; i++ {
-			p, err := types.NewProposal(TestProposal, proposalID, time.Now(), time.Now())
+			p, err := types.NewProposal(govgenhelpers.TestTextProposal, proposalID, time.Now(), time.Now())
 			suite.Require().NoError(err)
 
 			p.Status = s
@@ -123,6 +125,50 @@ func (suite *KeeperTestSuite) TestGetProposalsFiltered() {
 					suite.Require().Equal(tc.params.ProposalStatus, p.Status)
 				}
 			}
+		})
+	}
+}
+
+func (suite *KeeperTestSuite) TestGetVotingPeriod() {
+	tests := []struct {
+		name                 string
+		content              types.Content
+		expectedVotingPeriod time.Duration
+	}{
+		{
+			name:                 "text proposal",
+			content:              govgenhelpers.TestTextProposal,
+			expectedVotingPeriod: types.DefaultPeriodText,
+		},
+		{
+			name:                 "param changes proposal",
+			content:              govgenhelpers.TestParameterChangeProposal,
+			expectedVotingPeriod: types.DefaultPeriodParameterChange,
+		},
+		{
+			name:                 "software upgrade proposal",
+			content:              govgenhelpers.TestSoftwareUpgradeProposal,
+			expectedVotingPeriod: types.DefaultPeriodSoftwareUpgrade,
+		},
+		{
+			name:                 "cancel software upgrade proposal",
+			content:              govgenhelpers.TestCancelSoftwareUpgradeProposal,
+			expectedVotingPeriod: types.DefaultPeriodSoftwareUpgrade,
+		},
+		{
+			name: "unhandled proposal",
+			content: distrtypes.NewCommunityPoolSpendProposal(
+				"title", "desc", sdk.AccAddress{},
+				sdk.NewCoins(sdk.NewInt64Coin("igovgen", 1)),
+			),
+			expectedVotingPeriod: types.DefaultPeriod,
+		},
+	}
+	for _, tt := range tests {
+		suite.Run(tt.name, func() {
+			vp := suite.app.GovKeeper.GetVotingPeriod(suite.ctx, tt.content)
+
+			suite.Require().Equal(tt.expectedVotingPeriod, vp)
 		})
 	}
 }