+++ x/gov/keeper/querier_test.go
@@ -156,7 +156,7 @@ func TestQueries(t *testing.T) {
 	oneCoins := sdk.NewCoins(sdk.NewInt64Coin(sdk.DefaultBondDenom, 1))
 	consCoins := sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, app.StakingKeeper.TokensFromConsensusPower(ctx, 10)))
 
-	tp := TestProposal
+	tp := govgenhelpers.TestTextProposal
 
 	depositParams, _, _ := getQueriedParams(t, ctx, legacyQuerierCdc, querier)
 
@@ -169,8 +169,6 @@ func TestQueries(t *testing.T) {
 	_, err = app.GovKeeper.AddDeposit(ctx, deposit1.ProposalId, depositer1, deposit1.Amount)
 	require.NoError(t, err)
 
-	proposal1.TotalDeposit = proposal1.TotalDeposit.Add(deposit1.Amount...)
-
 	proposal2, err := app.GovKeeper.SubmitProposal(ctx, tp)
 	require.NoError(t, err)
 	deposit2 := types.NewDeposit(proposal2.ProposalId, TestAddrs[0], consCoins)
@@ -178,8 +176,6 @@ func TestQueries(t *testing.T) {
 	require.NoError(t, err)
 	_, err = app.GovKeeper.AddDeposit(ctx, deposit2.ProposalId, depositer2, deposit2.Amount)
 	require.NoError(t, err)
-
-	proposal2.TotalDeposit = proposal2.TotalDeposit.Add(deposit2.Amount...)
 
 	// TestAddrs[1] proposes (and deposits) on proposal #3
 	proposal3, err := app.GovKeeper.SubmitProposal(ctx, tp)
@@ -191,8 +187,6 @@ func TestQueries(t *testing.T) {
 	_, err = app.GovKeeper.AddDeposit(ctx, deposit3.ProposalId, depositer3, deposit3.Amount)
 	require.NoError(t, err)
 
-	proposal3.TotalDeposit = proposal3.TotalDeposit.Add(deposit3.Amount...)
-
 	// TestAddrs[1] deposits on proposals #2 & #3
 	deposit4 := types.NewDeposit(proposal2.ProposalId, TestAddrs[1], depositParams.MinDeposit)
 	depositer4, err := sdk.AccAddressFromBech32(deposit4.Depositor)
@@ -200,19 +194,12 @@ func TestQueries(t *testing.T) {
 	_, err = app.GovKeeper.AddDeposit(ctx, deposit4.ProposalId, depositer4, deposit4.Amount)
 	require.NoError(t, err)
 
-	proposal2.TotalDeposit = proposal2.TotalDeposit.Add(deposit4.Amount...)
-	proposal2.Status = types.StatusVotingPeriod
-	proposal2.VotingEndTime = proposal2.VotingEndTime.Add(types.DefaultPeriod)
-
 	deposit5 := types.NewDeposit(proposal3.ProposalId, TestAddrs[1], depositParams.MinDeposit)
 	depositer5, err := sdk.AccAddressFromBech32(deposit5.Depositor)
 	require.NoError(t, err)
 	_, err = app.GovKeeper.AddDeposit(ctx, deposit5.ProposalId, depositer5, deposit5.Amount)
 	require.NoError(t, err)
 
-	proposal3.TotalDeposit = proposal3.TotalDeposit.Add(deposit5.Amount...)
-	proposal3.Status = types.StatusVotingPeriod
-	proposal3.VotingEndTime = proposal3.VotingEndTime.Add(types.DefaultPeriod)
 	// total deposit of TestAddrs[1] on proposal #3 is worth the proposal deposit + individual deposit
 	deposit5.Amount = deposit5.Amount.Add(deposit3.Amount...)
 
@@ -239,6 +226,11 @@ func TestQueries(t *testing.T) {
 
 	deposit = getQueriedDeposit(t, ctx, legacyQuerierCdc, querier, proposal3.ProposalId, TestAddrs[1])
 	require.Equal(t, deposit5, deposit)
+
+	// Get updated proposals for assertions
+	proposal1, _ = app.GovKeeper.GetProposal(ctx, proposal1.ProposalId)
+	proposal2, _ = app.GovKeeper.GetProposal(ctx, proposal2.ProposalId)
+	proposal3, _ = app.GovKeeper.GetProposal(ctx, proposal3.ProposalId)
 
 	// Only proposal #1 should be in types.Deposit Period
 	proposals := getQueriedProposals(t, ctx, legacyQuerierCdc, querier, nil, nil, types.StatusDepositPeriod, 1, 0)